Add logo to the animated drawer

+++ lib/utils/typography.dart
@@ -1,6 +1,6 @@
 const List<int> fontWeight = const [300, 700];
 const List<double> fontScale = const [1, 1.2, 1.4, 1.6, 1.8, 2.0];
-const List<int> fontSize = const [
+const List<double> fontSize = const [
   10,
   12,
   14,
